feat(time-offset): allow fractional (sub-second) offsets
Time offsets were integer seconds, so sub-second clock skew couldn't be
corrected. Switch both devices.default_time_offset_sec and
videos.time_offset_sec to double precision seconds.

- migration 000016 alters the columns to double precision (down rounds back)
- regenerate sqlc: offset fields become float64
- Go handlers/DTOs and seed-dev use float64; effective time math uses
  fractional seconds
- OpenAPI: offset fields integer -> number; regenerate client
- offset NumberInputs gain step=0.1 and note decimals are allowed

+++ cmd/seed-dev/main.go
@@ -181,7 +181,7 @@ func ensureOwnTeam(ctx context.Context, q *sqlc.Queries, name string) (sqlc.Team
 	return q.CreateTeam(ctx, sqlc.CreateTeamParams{Name: name, IsOwn: true})
 }
 
-func ensureDevice(ctx context.Context, q *sqlc.Queries, name string, offset int32) (sqlc.Device, error) {
+func ensureDevice(ctx context.Context, q *sqlc.Queries, name string, offset float64) (sqlc.Device, error) {
 	devices, err := q.ListDevices(ctx)
 	if err != nil {
 		return sqlc.Device{}, err
